feat: Implement comprehensive offline synchronization and data management across multiple modules including Eau Minerale, Orange Money, Immobilier, Gaz, and Administration.

+++ lib/features/gaz/presentation/screens/gaz_shell_screen.dart
@@ -32,7 +32,7 @@ class _GazShellScreenState extends BaseModuleShellScreenState<GazShellScreen> {
     // Listen to the navigation provider to update the internal state
     ref.listen(gazNavigationIndexProvider, (previous, next) {
       if (next != selectedIndex) {
-        navigateToIndex(next as int);
+        navigateToIndex(next);
       }
     });
